+++ src/FPDivSqrt/SelFunctionTable.hpp
@@ -6,16 +6,20 @@
 
 namespace flopoco
 {
+
+	// dmin and dmax are 0.5 and 1 if there is no prescaling; In case of prescaling they can be larger.
+	// nbd and nbw are the number of bits of d and w to input to the selection table
+	// digit is alpha (the max absolute value of the digit)
 	class SelFunctionTable : public Table
 	{
 		public:
-			SelFunctionTable(Target* target, float dmin, float dmax, int nbd, int nbw, int digit, int base, int wIn, int wOut);
+			SelFunctionTable(Target* target, double dmin, double dmax, int nbd, int nbw, int digit, int base, int wIn, int wOut);
 			virtual ~SelFunctionTable();
 			mpz_class function(int x);
 		protected:
-			float Dmin;
-			float Dmax;
-			float ro;
+			double Dmin;
+			double Dmax;
+			double ro;
 			int nbBitD;
 			int nbBitW;
 			int digitSet; /**Must not exceed 10!!!**/
